Move to individual analysis rule.
